invader: split strongholds into mods/modern/stronghold
The NPC loop for userId 2 is re-registered from the stronghold mod, which
handles core rooms itself and delegates to the classic raid loop otherwise.
