+++ .changeset/invader-mod-split.md
@@ -1,0 +1,5 @@
+---
+"xxscreeps": patch
+---
+
+Split the invader mod into classic raids and modern strongholds.
