+++ packages/xxscreeps/mods/modern/stronghold/templates.ts
