+++ packages/xxscreeps/mods/modern/xx.d.ts
@@ -5,4 +5,5 @@ declare module 'xxscreeps:mods/constants' {
 	export * from 'xxscreeps/mods/modern/observer/constants.js';
 	export * from 'xxscreeps/mods/modern/powerbank/constants.js';
 	export * from 'xxscreeps/mods/modern/powerspawn/constants.js';
+	export * from 'xxscreeps/mods/modern/stronghold/constants.js';
 }
